remove unnecessary functions/variables

+++ Moon/RenderMesh.cpp
@@ -145,7 +145,7 @@ namespace Moon
 					fastgltf::Accessor& indexaccessor = gltf.accessors[p.indicesAccessor.value()];
 					fastgltf::iterateAccessor<std::uint32_t>(gltf, indexaccessor, [&](std::uint32_t idx)
 						{
-							indices.push_back(idx + initial_vtx);
+							indices.push_back(idx + (uint32_t)initial_vtx);
 						});
 				}
